requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.5.18.1
chardet==4.0.0
click==8.1.3
colorama==0.4.4
Cython==0.29.23
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjarowinkler==1.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
Sastrawi==1.0.1
scipy==1.7.3
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
nltk==3.7